+++ src/service/bookService.js
@@ -1,32 +1,33 @@
 const Book = require('../models/book')
 const Author = require('../models/author')
 const Genre = require('../models/genre')
+const BookinstanceService = require('./bookinstanceService')
 
 const async = require('async')
 
 exports.create = req => {
   return new Promise((resolve, reject) => {
-    const { title, author, summary, isbn, genre } = req.body
-    authorIds = [author._id]
-    genreIds = genre.map(g => g._id)
+    const {title, author, summary, isbn, genre} = req.body
+    const authorIds = [author._id]
+    const genreIds = genre.map(g => g._id)
     async.parallel(
       {
-        authors: function(callback) {
-          Author.find({ _id: [authorIds] }, callback)
+        authors: function (callback) {
+          Author.find({_id: [authorIds]}, callback)
         },
-        genres: function(callback) {
-          Genre.find({ _id: [genreIds] }, callback)
+        genres: function (callback) {
+          Genre.find({_id: {$in: [...genreIds]}}, callback)
         }
       },
-      function(err, results) {
+      function (err, results) {
         if (err) {
           return reject(err)
         }
         if (results.authors.length == 0) {
-          return reject({ msg: '作者不可用', author })
+          return reject({msg: '作者不可用', author})
         }
         if (results.genres.length == 0) {
-          return reject({ msg: '类型不可用', genre })
+          return reject({msg: '类型不可用', genre})
         }
         const book = new Book({
           title: title,
@@ -36,7 +37,7 @@ exports.create = req => {
           genre: typeof genre === 'undefined' ? [] : genre
         })
 
-        book.save(function(err) {
+        book.save(function (err) {
           if (err) {
             return reject(err)
           } else {
@@ -55,56 +56,61 @@ exports.delete = id => {
         return reject(err)
       }
       if (!result) {
-        return reject({ msg: '找不到该书籍', _id: id })
+        return reject(`找不到该书籍 ${id}`)
       }
-      Book.findByIdAndRemove(id, err => {
-        if (err) {
-          return reject(err)
+      BookinstanceService.count({book: result}).then(count => {
+        if (count > 0) {
+          return reject('该书籍已有库存，无法删除')
         }
-        return resolve({ id })
-      })
+        Book.findByIdAndRemove(id, err => {
+          if (err) {
+            return reject(err)
+          }
+          return resolve({id})
+        })
+      }).catch(err => reject(err))
     })
   })
 }
 
 exports.update = req => {
   return new Promise((resolve, reject) => {
-    const { _id, title, author, summary, isbn, genre } = req.body
+    const {_id, title, author, summary, isbn, genre} = req.body
     authorIds = [author._id]
     genreIds = genre.map(g => g._id)
     async.parallel(
       {
-        book: function(callback) {
+        book: function (callback) {
           Book.findById(_id).exec(callback)
         },
-        authors: function(callback) {
-          Author.find({ _id: [...authorIds] }, callback)
+        authors: function (callback) {
+          Author.find({_id: [...authorIds]}, callback)
         },
-        genres: function(callback) {
-          Genre.find({ _id: [...genreIds] }, callback)
+        genres: function (callback) {
+          Genre.find({_id: [...genreIds]}, callback)
         }
       },
-      function(err, results) {
+      function (err, results) {
         if (err) {
           return reject(err)
         }
         if (results.book == null) {
-          return reject({ msg: '找不到该书籍', _id })
+          return reject(`找不到该书籍 ${_id}`)
         }
         if (results.authors.length == 0) {
-          return reject({ msg: '作者不可用', author })
+          return reject({msg: '作者不可用', author})
         }
         if (results.genres.length == 0) {
-          return reject({ msg: '类型不可用', genre })
+          return reject({msg: '类型不可用', genre})
         }
-        const book = new Book({ _id, title, author, summary, isbn, genre })
+        const book = new Book({_id, title, author, summary, isbn, genre})
         console.log(book)
-        Book.findByIdAndUpdate(_id, book, {}, function(err, modifiedbook) {
+        Book.findByIdAndUpdate(_id, book, {}, function (err, modifiedbook) {
           if (err) {
             return reject(err)
           }
           if (modifiedbook == null) {
-            return reject({ msg: '找不到该书籍', _id })
+            return reject(`找不到该书籍 ${_id}`)
           } else {
             resolve(modifiedbook)
           }
@@ -124,7 +130,7 @@ exports.detail = id => {
           return reject(err)
         }
         if (result == null) {
-          return reject({ msg: '找不到该书籍', id })
+          return reject(`找不到该书籍 ${id}`)
         }
         resolve(result)
       })
@@ -133,13 +139,13 @@ exports.detail = id => {
 
 exports.list = query => {
   return new Promise((resolve, reject) => {
-    let { title, isbn, page, limit } = query
+    let {title, isbn, page, limit} = query
     page = page ? page : 1
     limit = limit ? limit : 10
     let query_ = {}
     let like = []
     if (title) {
-      like.push({ title: { $regex: new RegExp(title, 'i') } })
+      like.push({title: {$regex: new RegExp(title, 'i')}})
       query_.$or = like
     }
     if (isbn) {
@@ -154,11 +160,11 @@ exports.list = query => {
         if (err) {
           return reject(err)
         }
-        Book.count(query_, function(err, count) {
+        Book.count(query_, function (err, count) {
           if (err) {
             return reject(err)
           }
-          resolve({ items: result, total: count })
+          resolve({items: result, total: count})
         })
       })
   })
@@ -166,7 +172,7 @@ exports.list = query => {
 
 exports.count = query => {
   return new Promise((resolve, reject) => {
-    Book.count(query, function(err, count) {
+    Book.count(query, function (err, count) {
       if (err) {
         return reject(err)
       }
